apply_system_defaults: disable $Version check so that $$ values always over-ride
By default the variable "convert" will be set if $Version is found in
system_defaults and its value does not match the currently configured
value (the actual value isn't important, it just needs to be a
different string).

However, during development we want changes in system_defaults to be
applied without the need to repeatedly update $Version. Setting
ALWAYS_CONVERT effectively does that (ie it's the same as forcing the
$Version check to always detect a difference even if the values
match).

The end result should be that values in system_defaults defined with
$$ (e.g. $$dibbler_client_enable=true) will always over-ride any
setting which may have already been configured.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>