utopia_init.sh: drop support for legacy kernel versions
By default utopia_init.sh assumes that for kernel version >= 4.x the
netfilter control is in:

  /proc/sys/net/netfilter/

and otherwise (ie kernel version < 4.x) it's in:

  /proc/sys/net/ipv4/netfilter/

That logic doesn't work for kernel version 3.12.x, which uses the
same paths as kernel version 4.x.

Since the oldest kernel we expect to support is 3.12.74 (ie the
kernel in the Puma6 r6.1.8 SDK) we can simply remove the broken
version test rather than try to fix it (use the newer path
unconditionally).

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>