improved kernel version detection in utopia_init.sh
utopia_init.sh assumes that for kernel version >= 4.x then netfilter
control is in:

  /proc/sys/net/netfilter/

and otherwise (ie kernel version < 4.x) it's in:

  /proc/sys/net/ipv4/netfilter/

For the Puma6 r6.1.8 SDK, the kernel version is 3.12.74 but we have
netfilter control in /proc/sys/net/netfilter/ so the path is not
correctly detected.

  root@mv1intel-arm-systemd:~# uname -a
  Linux mv1intel-arm-systemd 3.12.74 #4 PREEMPT Fri Apr 3 05:16:45 UTC 2020 armv6b GNU/Linux

As a workaround update the kernel version check in utopia_init.sh to
use the /proc/sys/net/netfilter/ paths for kernel version > 3.11.

Note that the cut off point of 3.11 is set to catch our 3.12.x
kernel, not because it's known to be the correct cut off point in all
cases.

Testing for the existence of one of the kernel version specific paths
might be a simpler and reliable test?

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>