fix misuse of "\0"
"\0" is a pointer to an empty string (ie effectively the same as ""),
and is NOT the same as a nul charactor (ie a char with value 0).

Usage of "\0" is almost always a bug. If a pointer to an empty string
is needed then use "", if a nul character is needed then use 0 or
'\0'.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>