service_multinet: use fscanf() to parse psmcli output in nv_get_bridge() + misc cleanup
Using fscanf() is more efficient than the two step approach of using
fgets() to read the output into a temp buffer and then using sscanf()
to parse it. Using fscanf() also allows the 3 lines of output to be
parsed together (which isn't possible when using fgets() which stops
reading the input at each newline).

Apart from the improvement in efficiency, the main motivation for
this change is to improve the robustness of the parsing and avoid the
chance of crashing if the expected values are not found in psm (which
was seen in the previous version).

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>