Mercury V2 Use of public DNS servers ( CR039 RDKB-4591 ) ( system_defaults change only )
[Description]
- Added system default values for dns_config_page_show and dns_override

[Testing steps]:
- Factory reset modem
- Check default values for Device.X_LGI-COM_Gateway.DNS_Override and Device.UserInterface.X_LGI-COM_ShowDNSConfigPage

ensure that public DNS can only be enabled for customer ID 20 ( OFW-210 )

dns_override is enabled to true in lg_syscfg_cust_20.db
The same was removed from the genetic file (lg_syscfg_common.db). If some new customer has to enable Public DNS, it needs to be enablen int he project specific syscfg file.
Changes has been made in the ccsp-p-and-m module to check whether dns_override is enabled in syscfg. If dns_override is not set to true in syscfg the API will not set the param and further will not enable the Public DNS.

Testing Performed:
Tested by changing the Customer index to 20. Was able to set the dns_override
Tested to by setting to another Customer. Was not able to set the dns_override