ulog: fix multiple issues with ulog_GetProcId()
The local string buffers within ulog_GetProcId() were defined as
arrays of char * instead of arrays of char. The result of getpid()
(type pid_t) was assigned directly to "pid" (ie a pointer to pid_t),
etc.

It looks looks like the original version of this code would have
crashed at run time, so maybe it's not being called?

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>