use cmp (instead of diff) to check for file changes in service_dhcp_server.sh
Since there's no need to create a patch of any changes, usng cmp is a
more appropriate (and efficient) way to check for differences between
two files than diff.

Also cleanup and unify the two duplicate versions of the code in
dhcp_server_start() and restart_request().

Warning: cmp needs to be available at run time (e.g. enabled in
busybox - which it may not be by default).

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>