Support for IPv4 GRE Tunnel ( RDKB-4914 )
[Description]
Multiple gre vlan is created for a single bridge
Need to remove mutiple vlan creation for gretap and support without vlan, that is pvid set as 0.
This change is to make one gre vlan for all the interfaces connected on the bridge, if pvid is 0, no vlan will be created.

[Test Description]:
- Enable the hotspot
- 1 bridge, 2 ssid and 1 gre valn should be created.

Signed-off-by: Karthik Shanbhag(kshanbhag@libertyglobal.com)