execute_dir: drop unused support for builtin default directory
Always require the directory containing scripts etc to run to be
passed as the first argument on the execute_dir command line. All
callers of execute_dir within RDKB appear to do that already.

execute_dir is a basically a clone of the Debian run-parts utility
(provided by busybox etc). Our eventual goal is to switch to using
run-parts and drop execute_dir completely. The first step to doing
that is to make execute_dir behave just like run-parts (which
requires a directory to be passed on the command line).

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>