pmon: fix broken NULL pointer check in str_trim()
The argument passed to str_trim() is the address of a pointer (the
address of a local variable within main() and so can never be NULL).

However, the pointer itself may be NULL (e.g. in the case of a
malformed config file with missing options, by the time str_trim() is
called for the third time, strsep() may have set the pointer to NULL)
and therefore needs to be checked.

Fixes crashes seen in pmon under long term stress testing.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>