pmon.sh: fix variable name mixup in do_unregister()
The variable being set was FEATURE but the second sysevent set
command uses $feature. Update the variable name so that the two
match.

Even with the bug everything probably worked OK since
pmon_feature_$ct was being cleared (and that's checked first to
determine if an entry is in use).

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>