firewall: fix incorrect handling of rh0 packets ( OFW-311 )
Bug Description:
According to RFC6092 REC-4, packets that are forwarded from the LAN
with routing extension header type 0 should not be forwarded to the
WAN.

Additionally as per rfc4443#section-3.4 :
If an IPv6 node processing a packet finds a problem with a field in
the IPv6 header or extension headers such that it cannot complete
processing the packet, it MUST discard the packet and SHOULD
originate an ICMPv6 Parameter Problem message to the packet's source,
indicating the type and location of the problem.

It is observed that ICMP packets that are forwarded from the LAN with routing extension header type 0 are forwarded to WAN and ICMPv6 Parameter Problem packet is not sent to the LAN client.

FIX DESCRIPTION:
Packets are now discarded and a response is sent to LAN client with ICMPv6 Parameter Problem packet having correct code and pointer

Tests Performed:
1. Provision the gateway
2. Forward a packet with routing extension header type 0 from LAN to WAN
3. Verify that  the packet is not forwarded to WAN and ICMPv6 Parameter Problem packet with correct code and pointers is sent to LAN client.

Expected Result:
PASS: No ICMPv6 Echo Requests were forwarded through the gateway to the WAN
PASS: No ICMPv6 Echo Replies were returned through the gateway to the LAN
ICMPv6 Parameter Problem packet received with correct code and pointers set (ICMPv6 code is 0 '(erroneous header field encountered)' )