autotools cleanup ( don't add dirs above ${top_srcdir} to include paths )
Adding dirs such as:

  $(top_srcdir)/../CcspCommonLibrary/source/ccsp/include

to the include path is a leftover from the days when all CCSP
components were built together as part of a single package. It
doesn't make any sense now that that the CCSP components are built
separately (and now find each others header files via sysroot).

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>