service_dslite: fix interface ipip6tun0 not coming up in DSLite mode ( OFW-785 )
[Description]
When we enable DSLite mode (using config file) it does not work correcty after boot up (Tunnel ipip6tun0 is not up),
we need to do an extra reboot to get it working fine.

RCA:
In the case of factory reset or movement from the ipv4 to dslite mode, it takes time to receive the AFTR.
Before the AFTR is received, service dslite is triggered and execution ends with the below sysevent status.
Fix is added to wait till the AFTR is received

[Test Description]:
1. Device is IPv4 mode, change the boot config file to have IPv6-DSLite configurations.
2. Device has boot config file with IPv6-DSLite configurations.

In both the cases, DSLite mode must be working correctly with the ipip6tun0 tunnel created and with the correct endpoints.

Signed-off-by: Karthik Shanbhag(kshanbhag@libertyglobal.com)

fix DSLite is getting crashed with EndpointAssignmentPrecedence static ( OFW-1198 )

TESTING PERFORMED:
1.Bring up the modem.
2.Set the Device mode to ipv6Only.
3.Enable DSLite mode using the following commands:
  dmcli eRT setv Device.DSLite.InterfaceSetting.1.EndpointAddress string 2233:730::1
  dmcli eRT setv Device.DSLite.InterfaceSetting.1.EndpointAssignmentPrecedence string Static
  dmcli eRT setv Device.DSLite.InterfaceSetting.1.Enable bool true
  dmcli eRT setv Device.DSLite.Enable bool true
4.Check tunnel created properly
5.Connect LAN client and check LAN getting IP
6.ping from LAN client to wan servers like google.com using ping. Ping should work.

Signed-off-by: Arundhati <aghose.contractor@libertyglobal.com>

service_dslite: fix error checking for result from get_aftr()

The previous test for string of length 0 AND equal to "none" could
never be true. Change the logic from AND to OR.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>