service_dslite: fix interface ipip6tun0 not coming up in DSLite mode ( OFW-785 )
[Description]
When we enable DSLite mode (using config file) it does not work correcty after boot up (Tunnel ipip6tun0 is not up),
we need to do an extra reboot to get it working fine.

RCA:
In the case of factory reset or movement from the ipv4 to dslite mode, it takes time to receive the AFTR.
Before the AFTR is received, service dslite is triggered and execution ends with the below sysevent status.
Fix is added to wait till the AFTR is received

[Test Description]:
1. Device is IPv4 mode, change the boot config file to have IPv6-DSLite configurations.
2. Device has boot config file with IPv6-DSLite configurations.

In both the cases, DSLite mode must be working correctly with the ipip6tun0 tunnel created and with the correct endpoints.

Signed-off-by: Karthik Shanbhag(kshanbhag@libertyglobal.com)