add support for setting Customer index based on platform specific config ( OFW-570 )
When we upgrade mv1 to mng, we need to convert the existing operator id to customer index. Since many countries are using the same operator id, it is not unique.
But country id is unique in mv1. So during the first upgrade or a factory reset, read the country id from CbnDocsisDb and write it in syscfg.
Also made changes in syscfg to handle customer index change via dmcli/boot-cofig-file without factory reset.

The value of customer-index-changed is changed from True to true so the extra check can be removed.

Testing Performed:
1. Factory reset the DUT
2. Device.X_LGI-COM_General.CustomerId should contain a value other than zero
3. Change Device.X_LGI-COM_General.CustomerId via dmcli
4. syscfg and psm db values should change according to the new customer index

Signed-off-by: Cibil <cpankiras.contractor@libertyglobal.com>

syscfg: only set "customer-index-changed" to true corresponding customerId specific file exists

Prevent changes to /nvram/bbhm_cur_cfg.xml getting lost after a
reboot when customerId is 0.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>

Map OperatorId to Customer index ( OFW-570 )

Moved the logic of converting CountryId to Customer Index to platform hal

Signed-off-by: Cibil <cpankiras.contractor@libertyglobal.com>

syscfg: allow Customer Index to be set to 0 ( OFW-1004 )

If not able to retrieve the customer index from hal or not able to read the customer-specific default file
then set the Customer_Index as zero. Also if Customer index was changed via dml then set the Customer_Index as zero

Signed-off-by: Cibil <cpankiras.contractor@libertyglobal.com>