fix ipip6tun0 not coming up in DSLite mode ( OFW-785 OFW-672 )
[OFW-785] : Interface ipip6tun0 not coming up in DSLite mode
[OFW-672] : IPv4 traffic disruption is seen in DSLite mode when the eRouter IP changes

[Description]
When we enable DSLite mode (using config file) it does not work correctly after boot up (Tunnel ipip6tun0 is not up),
we need to do an extra reboot to get it working fine.

wait added in the previous commit is removed and made a event based mechanism
One more corner case where "zebra start" was hung, as the lock is acquired by the service_dslite, it was a dead lock situation
For this reason, "zebra start" and "Zebra stop" are removed
Instead, killall zebra and sysevent set zebra-restart is used.

[Test Description]:
1. Device is IPv4 mode, change the boot config file to have IPv6-DSLite configurations.
2. Device has boot config file with IPv6-DSLite configurations.

In both the cases, DSLite mode must be working correctly with the ipip6tun0 tunnel created and with the correct endpoints.

Signed-off-by: Karthik Shanbhag(kshanbhag@libertyglobal.com)