utopia_init_xb6.sh: more robust parsing of imagename from /version.txt
The eventual goal is to be able to source /version.txt to extract the
image name (instead of needing to parse it with grep, sed, etc). As a
transitional step towards that, update parsing to support all
combinations of : or = as separator and string quoted or not quoted.
ie /version.txt containing any of:

 1: imagename:mng-image-rdkb-puma6-arm-qemuarm-20210326015040
 2: imagename=mng-image-rdkb-puma6-arm-qemuarm-20210326015040
 3: imagename:"mng-image-rdkb-puma6-arm-qemuarm-20210326015040"
 4: imagename="mng-image-rdkb-puma6-arm-qemuarm-20210326015040"

Currently /version.txt contains 1), but will eventually become 4).

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>