fix following LAN Traffic Test Stability Result Analysis ( OFW-1152 ) - Puma6 ARM specific
[Description]
ddh getting enabled, and then the packets start dropping

[RCA]
brlan0 is was not in the white list of the packet processor.
only l2sd0 is available. for the same reason packets on brlan0 starts dropping.

This fix makes brlanX mac address same as that of l2sd0, so that all the brlanX interfaces are in white list.

Signed-off-by: Karthik Shanbhag(kshanbhag@libertyglobal.com)