syscfg: drop support for syscfg in a raw MTD device
There are two options for syscfg persistent storage:

  Direct write to an MTD partition (ie NOR Flash)
  Write to filesystem (ie NAND Flash via filesystem and FTL)

There are no current (or even legacy) RDKB platforms which store
syscfg data in a raw MTD partition. The code to do so is therefore
unused, untested and unmaintained. Remove it.

Drop obsolete references to syscfg in a raw MTD partition from
utopia_init.sh.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>