Revert "optimize syscfg calls from scripts ( OFW-1153 )"
The way sysevents are handled (the split between shell scripts and C
code) is not consistent between platforms. For example see
utopia/source/scripts/init/c_registration/15_dhcp_server.c:

  const char* SERVICE_DEFAULT_HANDLER = "/etc/utopia/service.d/service_dhcp_server.sh";

  #if defined(_COSA_INTEL_USG_ARM_) && !defined(INTEL_PUMA7) && !defined(_COSA_BCM_ARM_) && !defined(_PLATFORM_IPQ_)

  const char* SERVICE_CUSTOM_EVENTS[] = {
      "syslog-status|/etc/utopia/service.d/service_dhcp_server.sh",
      "lan-status|/usr/bin/service_dhcp",
      "dhcp_server-restart|/usr/bin/service_dhcp",
      "dhcp_server-start|/usr/bin/service_dhcp",
      "dhcp_server-stop|/usr/bin/service_dhcp",
      "dhcp_server-resync|/etc/utopia/service.d/service_dhcp_server.sh|NULL|"TUPLE_FLAG_EVENT,
      NULL};

  #else

  const char* SERVICE_CUSTOM_EVENTS[] = {
      "syslog-status|/etc/utopia/service.d/service_dhcp_server.sh",
      "lan-status|/etc/utopia/service.d/service_dhcp_server.sh",
      "dhcp_server-resync|/etc/utopia/service.d/service_dhcp_server.sh|NULL|"TUPLE_FLAG_EVENT,
      NULL};

  #endif

In this example, _COSA_INTEL_USG_ARM_ targets use "service_dhcp" (ie
a binary) to handle some sysevents but other targets use only the
"service_dhcp_server.sh" shell script.

Therefore the recent optimisations to the service scripts which may
be safe when the script is only used to handle certain sysevents are
not safe if the script is used to handle all sysevents.

The long term solution is to make the split between shell scripts and
C code consistent across all platforms and then fully optimise the
scripts for sysevents which they still need to handle. However, in
the short term revert the recent optimisations to the shell scripts.

This reverts commit 4df399d0096137511d57ab5538a2d784196d8078.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>