syscfg: add new set of syscfg_set() APIs
Add a range of new APIs to set string or integer values, with or
without a namespace and with or without a call to syscfg_commit().

There is a lot of duplicated code within RDKB to do common tasks such
as setting a syscfg value from an integer or setting a syscfg value
followed by calling syscfg_commit(). RDKB can be simplified (ie made
smaller, easier to understand and less likely to contain bugs) by
providing APIs to perform these common tasks directly.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>