syscfg: don't report errors from syscfg_commit()
The purpose of syscfg_commit() is to write a copy of the current
state of syscfg to a file in the filesystem. There is very little
that application code can do if that fails other than output debug.
Outputting debug from all the places which call syscfg_commit() is
not very efficient (if debug output is needed it would be better to
generate it directly from within syscfg_commit()).

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>