syscfg: drop support for "syscfg destroy" ( use syscfg_destroy command instead )
Previously there were two ways in invoke syscfg destroy from a shell
prompt or schell script: by calling the "syscfg" command with
"destroy" as the command line option or by calling the
"syscfg_destroy" command (which is a symlink to "syscfg").

All scripts etc in RKDB are calling "syscfg_destroy", so we can drop
support for handling destroy as a command line option to syscfg.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>