service_dhcpv6_client_arm.sh: update DHCPV6_PID_FILE path to dibbler .pid file
The default path is /var/lib/dibbler/client.pid but for unknown
reasons service_dhcpv6_client_arm.sh has been expecting
/tmp/dibbler/client.pid and relying a symlink from one to the other.

The symlink is created by the meta-mng version of dibbler-init.sh but
it may not be created if BSP layers are providing their own version
of dibbler-init.sh. Safest option seems to be to update
service_dhcpv6_client_arm.sh to use the default .pid file path and
stop relying on the symlink.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>