service_routed: use ra_interval from syscfg instead of hardcoded default
Fetch ra_interval from syscfg for most target (previously it was only
done for Puma 7).

Also improve robustness of check for DNSSL Lifetime:

The DNS Search List Lifetime should be at least 3 * MaxRtrAdvInterval

In most cases, the value of MaxRtrAdvInterval is fetched from the
"ra_interval" value in syscfg. If a value is not found in syscfg then
the DNSSL Lifetime check was using atoi() of the empty buffer
returned from syscfg_get().

Update the code so that if a default value is used for ra_interval
then that default value is also used to check DNSSL Lifetime.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>