service_dhcp_server: add --dhcp-authoritative to dnsmasq command line in all cases
Historically dnsmasq was sometimes started with --dhcp-authoritative
and sometimes without. That was changed on 2020-11-04 by the Comcast
commit:

  https://github.com/lgirdk/utopia/commit/faa8d307bb822b8600b1a3d73509610d89752345

This commit replaced various individual calls which started dnsmasq
(some with --dhcp-authoritative and some without) with a single
common function which did not pass --dhcp-authoritative (ie this
change disabled --dhcp-authoritative in all cases).

It's not clear why dnsmasq was originally sometimes started with
--dhcp-authoritative and sometimes without but it's required in at
least some cases.

Update the code to pass --dhcp-authoritative in all cases.

If there are cases where dnsmasq should be started without
--dhcp-authoritative then further updates to the code will be
required.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>