apply_system_defaults: minor code cleanup in parse_line()
If a leading control character is found, then start parsing from the
next character, rather than relying on parse_line() to always skip
the first character of a line. This makes parse_line() more generic
and allows it to also be used to parse lines which don't begin with
a leading control character.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>