drop utctx_cmd SYSCFG_FAILED error handling
If utctx_cmd sets SYSCFG_FAILED the it either means utctx_cmd was
called with invalid command line options or syscfg failed (ie it's
not a routine error case such as syscfg value not found etc).

Trying to handle SYSCFG_FAILED complicates future updates to remove
the need for "eval" with utctx_cmd (and there's not much a script
calling utctx_cmd can do in response anyway) so simplify the code and
drop attempts to handle it.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>