enable Mv2+ specific iptable for bridged traffic ( OFW-2176 )
[Issue]
Iptable for bridges were not enabled in MV2+, for this dhcp option is not getting inserted in the community wifi
BCM response was, enabling will have affect on hw acceleration
A kernel patch was added to handle only bootp packets, so that without enabling iptable for bridge dhcp option can be inserted.

[Description]
After the kernel patch ebtable was required to add the marking to the packets

[Solution]
Added the ebtable rule, if hotspot is enabled.
And add the value to proc entry

[Testing Performed]
Enable Hotspot, connect the client capture packets on gretap0, confirm dhcp option are inserted correctly.

Signed-off-by: Karthik Shanbhag(kshanbhag@libertyglobal.com)