utopia_init.sh: don't unset UpdateNvram syscfg variable
The UpdateNvram syscfg variable is obsolete and doesn't control
anything anymore. However, if it's already in syscfg then leave it
there to avoid potential issues if a system is downgraded to an older
build which does use it.

Even if the value is definitely not going to be needed, removing it
isn't worth the overhead of calling syscfg commit during every boot!

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>