utopia_init.sh: increase priority of rpcserver on Puma6 ARM ( OFW-2283 CH7465SI-153 )
Calls to rpcclient from Puma6 Atom to ARM during boot are slow (and
sometimes fail completely) due to timeouts due to high CPU load on
the ARM. This slows the Atom boot process and/or requires Atom init
scripts etc to retry calls to rpcclient if they fail.

Increase the priority of rpcserver on Puma6 ARM solves these issues
and gives an overall reduction in the boot time for Puma 6 platforms.

Also fix debug so that "Starting rpcserver ..." is only shown when
rpcserver is actually being started.

Signed-off-by: Mustafa Karadzha(mkaradzha@libertyglobal.com)