service_udhcpc: add support for decoding ACS URL and proivisioning code from DHCPv4 option 43 ( OFW-2566 F5685SI-19 )
[Issue Description]

Implementation for requirement , The gateway firmware MUST support to acquire the ManagementServer URL (Device.ManagementServer.URL) from the DHCP server in the Vendor Specific Information - DHCPv4 option 43

[fix Description]

The implementation is to parse the option 43.1 for ACS URL and 43.2 for corresponding provisioning code to reflect in Device.DeviceInfo.ProvisioningCode.

[Testing Performed]

Verified that the device comes online with the ACS URL received from DHCPv4 option 43.

Signed-off-by: BijeeshaVS <bvattakaithayilsudha.contractor@libertyglobal.com>

service_udhcpc: backoff mechanism for DHCPDISCOVER when option 43 was not part of DHCPOFFER

[Description]
DHCPv4 option 43 and DHCPv4 option 125 contains ACS URL. Mv3 platforms (Ethan and Bosa)
need to get ACS URL from DHCPv4/6 for initial provisioning(for firmware upgrade and data model
configurations).

[Solution]
As per MVXREQ-2849, If option 43 was not received in DHCPv4 offer, then the gateway
should retransmit DHCPDISCOVER message. As a fix, check for ACS URL in DHCP offer,
and if it is not available then return to retry DHCP discover.

[Testing Performed]
1. Provision the gateway in IPv4 mode.
2. Connect the gateway to a DHCPv4 server that doesn't provide option 43
3. Capture the packets on erotuer0
4. DHCP client (udhcpc) should send DHCPDISCOVER packets as per
the backoff mechanism mentioned on MVXREQ-1150

Upstream Candidate: No. LGI-specific DHCP discover retry mechanism. MVXREQ-2849, MVXREQ-1150

Note,
service_udhcpc is only applicable for Mv3. If we port service_udhcpc to other platforms
then this change should be a patch applied from an Mv3 specific meta layer.

Signed-off-by: Saikiran Vanamala <svanamala@libertyglobal.com>