service_wan: update WAN status when WAN is stopped in dslite mode ( OFW-2703 )
BUG DESCRIPTION:
DHCPv4 lease time not updated with 120 second while RF disconnected in DsLite mode

ROOT CAUSE:
In dslite mode, when wan is stopped, the wan status is not updated.
The DHCPV4 lease time will be set to 120 depending on the wan-status.
As wan status was not updating the lease time remained unchanged.

FIX DESCRIPTION:
Updated the wan status on dslite mode.

TESTING PERFORMED:
1. Loaded the DUT with current firmware
2. Connect a LAN client and ensure the lease time is as 86400 on UI and on the client side
3. Disconnect RF and check if the DUT is as Offline in the CMTS
4. Disconnect and reconnect the LAN Client and check if the lease is now changed to 120 sec on the client side and on the UI (Under DHCP page)

Signed-off-by: Sreeja S Kumar (srekumar.contractor@libertyglobal.com)