utopia_init.sh: factory reset code clean up ( OFW-2340 )
[BUG DESCRIPTION]
Need to revert the temporary fix done in OFW-2336

Issue in OFW-2336:
Factory reset from UI/dmcli is not changing the channel config to
defaults because the device reboots while the wifi side factory reset
is pending for execution.

[FIX DESCRIPTION]
The values of FactoryResetSSID are read from the psm config backup
xml file and according to that wifi_factory_reset_ssid syscfg it set, to
handle the WiFi side Factory Reset during the next boot up if the
wifi factory reset is incomplete.

[TESTING PERFORMED]
Factory reset the device and verified the wifi_factory_reset_ssid
is set to 1 only if the WiFi Factory Reset is incomplete. When the device
is up after the Factory Reset, enabled the UI and checked the channel configs,
wifi SSID and password. All are set to defaults.

Upstream Candidate: No. LGI specific fix.
See the related commits in utopia and ccsp-p-and-m

Signed-off-by: Akash Dhanabal <adhanabal.ext@libertyglobal.com>