drop contents from dummy autoconf.h
None of the CONFIG_XXX options defined in the dummy autoconf.h are
used by RDKB code, so there's no need to define them. The SOC SDK may
still provide it's own autoconf.h (ie the kernel config) but it's a
clean solution and the long term goal should be that RDKB code does
not have any direct build time dependency on the kernel config.

Signed-off-by: Andre McCurdy <armccurdy@gmail.com>