service_wan: fix eRouter doesn't get IPv4 address Switching from bridge-mode to router-mode ( F3896SI-1208 OFW-3090 )
[DESCRIPTION]
During wan boot up if erouter0 interface if not ready and udchpc is not sending dhcp request in response to dhcp offer.

[FIX DESCRIPTION]
Introduced a wait for the interface to come up during wan start.

[TESTING PERFORMED]
verfied erouter0 status during mode switch and normal bootups.

Fixme:
Needs to be revisited to check the possiblities of handling the scenario in udhcpc handler.

Upstream Candidate: No. temporary fix

Signed-off-by: BijeeshaVS <bvattakaithayilsudha.contractor@libertyglobal.com>