fix ensure that previous instances are killed and tentative IPv6 addresses are progressed from tentative state ( OFW-2913 )
[Issue]
mcproxy fails to open the MRT socket if the already running instance is not terminated fully.
It is observed that killall command is not effective immediately.
Other isue is also mcproxy fails to initialize if the downstream interface has a tentative
IPv6 address at the time of initialization. To reproduce the issue:

reassign_local_ip() {
  local ip=$(ip a l dev $1 | awk '$1=="inet6" && $2 ~ /^fe80/ { print $2 }')
  ip a d $ip dev $1
  ip a a $ip dev $1
}

reassign_local_ip brlan0; sysevent set mldproxy-restart

[Solution]
Wait for previous mcproxy instances are not running before starting it.
Also wait for the end of IPv6 tentative state.

Signed-off-by: Mustafa Karadzha(mkaradzha@libertyglobal.com)