system_defaults_arm: update default values for multicast_max_ssm_sessions & multicast_m2u_max_sessions ( OFW-3309 )
Support to configure maximum number of M2U forwarding sessions via ACS and change value to 64 by default

[Description]
The DM is not exposed to the ACS and the current value for the M2UMaxSessions is 4.
Also update the MaxSSMSessions to 64 to make it redundant. The idea is to eventually
deprecate the parameter.

[Solution]
Expose the DM to the ACS and change the defaults for M2UMaxSessions and MaxSSMSessions to 64.

[Testing Performed]
1. Boot up the gateway
2. Once the device is online on the ACS, query the Device.X_LGI-COM_Multicast. DM.
3. Ensure that it following DMs are working -

        Device.X_LGI-COM_Multicast.Enable
        Device.X_LGI-COM_Multicast.SnoopingEnable
        Device.X_LGI-COM_Multicast.IGMPv3ProxyEnable
        Device.X_LGI-COM_Multicast.MLDv2ProxyEnable
        Device.X_LGI-COM_Multicast.SSMForwardingEnable
        Device.X_LGI-COM_Multicast.MaxSSMSessions
        Device.X_LGI-COM_Multicast.M2UMaxSessions
        Device.X_LGI-COM_Multicast.MaxThrottlingRate
        Device.X_LGI-COM_Multicast.MaxThrottlingHoldTime
        Device.X_LGI-COM_Multicast.FastLeaveEnable

4. Also verify that the value of M2UMaxSessions and MaxSSMSessions is 64 by default.

Upstream Candidate: No. LGI specific fix

Signed-off-by: Ajay Vishnu Bhat <abhat@libertyglobal.com>