firewall: enable the dos attack iptable rules to all OFW platforms ( OFW-3277 F5685SI-555 )
[Issue]
Unit is susceptible to a DNS Amplification attack.

[Description]
After adding the iptable rules ,we see the attack from DNS amplification LAN1 to LAN2 is not seen anymore.
But the CPU load reaches to 100% because of broadcom processes. This is yet to handled.

[Solution]
NA

[Testing]
Connect 2 LAN clients via ETHERNET cable behind the CPE.
One LAN client is the attacker and the other LAN client is the victim.
From attacker PC flood the DNS requests with source IP as victims IP address.
Then we see the CPE cpu load is 100% dnsmasq process is consuming 100% load.
Also, we see the victim PC is flooded with DNS responses and PC becomes not responsive for anything.

With the fix, we see dnsmasq reaching 100% load is fixed.
Also, the victim LAN client is safe. However, brm process still consumes 100% cpu load.

Upstream Candidate: No. LGI specific fix

Signed-off-by: Sidda(sihiregoudar@libertyglobal.com)