service_dhcp: use ipv6 wan dns server in case ipv4 dns server is not available ( OFW-3341 CH7465SI-436 )
In DSLite mode DUT fails to provide the default LG IPv4 & IPv6 DNS servers when customer configured IPv4 & IPv6 Preferred DNS are unreachable.

Description:
When customer configures IPv4 & IPv6 Preferred DNS as Valid(not reachable), the default DNS Server addresses are not provided to the client by CPE in the DHCP messages.
So, default LG provided IPv4 & IPv6 DNS servers are available during this scenario and thus clients fail to access internet.

Solution:
When the device is in ipv6/ipv6 Ds-lite mode, it is unable to fetch the defalut ipv4 DNS and thus fails to provide a reachable DNS Server addresses to the client. So, in this scenario we can fetch the default ipv6 DNS server address instead.

Testing Performed:
1) Switch to  customer defined DNS servers option in the UI and Provide any valid not reachable public DNS server IP addresses on IPv4 & IPv6 Preferred fields,
leave the Alternate fields of IPv4 and IPv6 as empty and apply changes in the UI.
2) Connect a Lan client or perform release renew on the exsiting client.
3) Query any domain name to verify that the dns query is sent using the default DNS Servers
4) root@modem-2-3-lan1:~# nslookup -q=A acs_server.boardfarm.com
   Server:		192.168.178.1
   Address:	        192.168.178.1#53

   Name:	acs_server.boardfarm.com
   Address:      172.25.2.40
   Name:	acs_server.boardfarm.com
   Address:      172.25.1.40

5) root@modem-2-3-lan1:~# nslookup -q=AAAA acs_server.boardfarm.com
   Server:		192.168.178.1
   Address:	        192.168.178.1#53

   Name:	acs_server.boardfarm.com
   Address:     2001:dead:beef:3::40
   Name:	acs_server.boardfarm.com
   Address:     2001:dead:beef:2::40

Now, the DUT provides the default DNS server to the client.

Upstream Candidate: Maybe, On top of public DNS feature.

Signed-off-by: mjothimani.ext <mjothimani.ext@libertyglobal.com>