service_ntpd.sh: add support for ntp_minpoll and ntp_maxpoll ( OFW-3789 )
max interval poll time should be 24 hours.

[Issue]
NTP was sending poll requests based on the defaults. 2^6 and 2^10

[Description]
This issue was discussed in CCB and NTP should be sending the polls between 2^10 and 2^16 by default.
And need datmodels to control the min and max polling interval

[Solution]
Add polling interval configurations to ntp conf.

[Testing]
Load the image, Confirm NTP is sending the the polls in specified interval in a incremental way.

Note: This commit includes the default configuration, Data model support
yet to be implemented.

Signed-off-by: Karthik Shanbhag(kshanbhag@libertyglobal.com)