add support for MinPollInterval & MaxPollInterval data models ( OFW-3789 )
Part of fix NTP - Max interval poll time should be 24 hours.

[Issue]
NTP was sending poll requests based on the defaults. 2^6 and 2^10

[Description]
This issue was discussed in CCB and NTP should be sending the polls between 2^10 and 2^16 by default.
And need datmodels to control the min and max polling interval

[Solution]
Implement Polling interval and datamodels as per MVXREQ-3867

[Testing]
Load the image, Confirm NTP is sending the the polls in specified interval in a incremental way.

Upstream Candidate: Maybe, New Addition

Signed-off-by: Karthik Shanbhag(kshanbhag@libertyglobal.com)