firewall: IPv6 multicast forwarding and acceleration fixes ( OFW-3885 )
[Issue]
IPv6 multicast forwarding is blocked in firewall. Also bypassing the firewall
shows that packets are not being accelerated.

[Solution]
The problem with filtering is due to erouter driver not setting the correct
pkt_type value and all of them are marked as unicast. Adding IP destiantion
address checks in firewall to handle incorrect values.

The acceleration issue was only seen in UDP port 5060 which we have a
netfilter helper registered for it (nf_conntrack_sip). As we do not need
any netfilter alg helpers for mABR streams, we can disable connection tracking
to avoid this issue for mABR SSM multicast address range.

Upstream Candidate: No. mABR is LGI specific

Signed-off-by: Mustafa Karadzha(mkaradzha@libertyglobal.com)